% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_souporcell_seurat}
\alias{add_souporcell_seurat}
\title{Add souporcell clustering file output to a seurat object}
\usage{
add_souporcell_seurat(
  obj,
  souporcell_file,
  prefix = NULL,
  assay_name = "GENO",
  key = "gpca_",
  meta_data_label = "geno",
  rd_label = "gpca",
  rename_assignments = T
)
}
\arguments{
\item{obj}{input Seurat object}

\item{souporcell_file}{location of souporcell clusters.tsv file}

\item{prefix}{optional prefix to prepend cell barcodes}

\item{assay_name}{name of assay object}

\item{key}{name of key to use}

\item{meta_data_label}{name of the column in the returned meta data to place the souporcell assignment.  Does not check for overwriting}

\item{rd_label}{name of the dimensionality reduction}

\item{rename_assignment}{rename souporcell assignment as discussed in the description}
}
\value{
a seurat object with a reduced dimension assay containing the SNP clusters as generated by souporcell but modified as discussed above.
A column is added to the meta data as designated by the meta_data_label argument with the genotype assignment generated by souporcell
}
\description{
This function takes as input a seurat object and the location of a souporcell file (clusters.tsv).  The function will add
an assay to the seurat object such that the reduced dimensions of said assay are the principal components of the log of the cluster probabilities
returned by souporcell in the clusters.tsv file.  These cluster probabalities are normalized by the mean before log transformation and PCA.
The souporcell 'assignment' is added to the seurat metadata using the meta_data_label argument.  Name of the dimensionality reduction can be set
using the rd_label argument, accompanying key is set with the key argument and the assay name is set using assay_name.  Use of the rename_assignments
parameter will set the genotype numbering to 1-indexed (instead of the 0-indexed souporcell output) and will collapse all genotype 
multiplet designations (0/1, 1/0, 0/2, 2/0, etc) to "multiplet".  This feature is turned off by setting rename_assignments to F.
}
