% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{augment_data}
\alias{augment_data}
\title{Aubment data}
\usage{
augment_data(obj, column, norm_number = 2000, assay = "RNA")
}
\arguments{
\item{column}{column from the metadata that designates cell group (i.e. celltype)}

\item{norm_number}{cell number to augment data to for cells that are not sufficiently abundant in the}
}
\value{
a seurat object augmented with simulated cells such that all cell groups are present at a level of norm_number of cells
}
\description{
This function takes a seurat object and finds cells that are not sufficiently abundant when grouped by the
column parameter, then simulates data to augment cell number to a level of the parameter - norm_number
}
