% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{keras_helper}
\alias{keras_helper}
\title{Keras helper}
\usage{
keras_helper(
  x_train,
  x_test,
  y_train,
  y_test,
  num_classes,
  query,
  learning_rate,
  verbose,
  layers = layers,
  max_epochs = max_epochs,
  batch_size = batch_size,
  device = device,
  keras_model = NULL
)
}
\value{
model evaluation of query
}
\description{
A function for input of viewmastR data into keras for training and evaluation of a query
}
