% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{lasso_helper}
\alias{lasso_helper}
\title{lasso helper}
\usage{
lasso_helper(
  x_train,
  x_test,
  y_train,
  y_test,
  num_classes,
  query,
  learning_rate,
  verbose,
  device,
  argg
)
}
\value{
model evaluation of query
}
\description{
A function for input of viewmastR data into lasso for training and evaluation of a query
}
