% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{smr_sparse_helper}
\alias{smr_sparse_helper}
\title{softmax regression helper for CSR sparse matrices}
\usage{
smr_sparse_helper(
  Xtrain,
  Xtest,
  Ytrain,
  Ytest,
  length_labels,
  query,
  learning_rate,
  verbose,
  device,
  iterations,
  lambda,
  max_error
)
}
\value{
model evaluation of query
}
\description{
A function for input of viewmastR data into keras for training and evaluation of a query
}
