\name{viewmastR2-package}
\alias{viewmastR2-package}
\alias{viewmastR2}
\docType{package}
\title{
\packageTitle{viewmastR2}
}
\description{
\packageDescription{viewmastR2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{viewmastR2}
\packageIndices{viewmastR2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{viewmastR2}

Maintainer: \packageMaintainer{viewmastR2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
