% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{vm_demo}
\alias{vm_demo}
\title{demo}
\usage{
vm_demo(
  perc = 80,
  FUNC = c("naive_bayes", "neural_network", "softmax_regression", "deep_belief_nn",
    "logistic_regression", "bagging", "perceptron", "keras"),
  device = c("GPU", "GPU_OPENCL", "GPU_CUDA", "CPU")
)
}
\arguments{
\item{perc}{percent to holdout}

\item{FUNC}{algorithm to use}
}
\description{
run classification demo
}
