% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{xgboost_helper}
\alias{xgboost_helper}
\title{xgboost helper}
\usage{
xgboost_helper(
  x_train,
  x_test,
  y_train,
  y_test,
  num_classes,
  query,
  verbose,
  argg
)
}
\value{
model evaluation of query
}
\description{
A function for input of viewmastR data into xgboost for training and evaluation of a query
}
